+++ CHANGELOG.md
@@ -6,12 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- `Select` - removed value from internal state
+- `Autocomplete` - removed defaultValue from internal state
+
 
 ## v6.0.6 - 2021-10-08
 
 ### Fixed
 - `Select` - fixed defaultValue not working
-- `Icon` Fixed document reference
+- `Icon` - fixed document reference
 
 
 ## v6.0.5 - 2021-10-07